fix(archtest): 消息加 INVARIANT ID + typeOwner 迁 helpers_test + 索引式 reflect 盲区 backlog (#948)
回应 review 后续三项：

#2 CI 可诊断性：7 处 reflect blind-spot per-violation 消息补 INVARIANT ID 前缀
   （session / credential_authority×2 / domain_authz×2 / caching /
   credential_invalidate），与 reconstitute/healthz/outbox 已有 ID 的格式统一；
   每条 CI 失败行现可单独 triage。

#4 typeOwner（3 caller：credential / session / reflect_string_arg）从
   credential_authority_assert_funnel_test.go 迁到 helpers_test.go 共享 helper
   归属处，消除跨文件 ghost 依赖。

#3 索引式 reflect 字段/方法读（Field(i)/FieldByIndex/FieldByIndexErr/Method(i)，
   无 string 实参）——结构上超出 string-arg scanner 范围，关闭需 struct 类型 +
   索引→名解析 + dataflow + 高误报面，部分覆盖会给虚假闭合感。开 backlog #1120
   跟踪（含两个设计选项），godoc out-of-scope 盲区段引用 #1120。
